Move widgets funcitons to launch Widget
